Added to docker compose and created a test script for the api script
There is now a user-defined network defined in the docker-compose file. This allows the containers to communicate with each other. Also, the API container is now created. When docker-compose is run, it will create all three containers and run a test script that will ensure the API container can connect to the database container and retrieve data from it.